Rename module :doc-parser to :code-gen
