--- doc-parser/build.gradle.kts
@@ -1,23 +1,0 @@
-plugins {
-    alias(libs.plugins.kotlin.jvm)
-    alias(libs.plugins.kotlin.serialization)
-    alias(libs.plugins.jetbrains.compose)
-    alias(libs.plugins.compose.compiler)
-}
-
-group = "com.gabrieldrn.docparser"
-version = "unspecified"
-
-dependencies {
-    implementation(libs.kotlinx.serialization.json)
-    implementation(kotlin("reflect"))
-    implementation("org.jetbrains.compose.ui:ui-graphics:${libs.versions.compose.get()}")
-    implementation(compose.runtime)
-    implementation(libs.kotlinPoet)
-
-    testImplementation(kotlin("test"))
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
